Move workspace package export files
